working on EventInventoryJob; added triggers for ProduceableJob; fixed issue with Part and Material can_be_produced (was decimal, now float, coerced to int); CountTransferJob: strange issue with item.set_history_from_current_counts(), might have been typo?; Assembly: remove trial code; Itemable: tweak callbacks; Count: remove #not_components scope; Part: alias #orderable and #not_made_from_material; RailsAdmin: fix material has_many parts joins, display Item#can_be_produced on show and edit.

+++ db/migrate/20210620175809_create_assemblies_join_table.rb
@@ -23,6 +23,6 @@ class CreateAssembliesJoinTable < ActiveRecord::Migration[6.1]
     # Part belongs_to Material
     add_reference :parts, :material
     rename_column :parts, :made_from_materials, :made_from_material
-    add_column :parts, :quantity_from_material, :decimal, precision: 8, scale: 4, default: nil
+    add_column :parts, :quantity_from_material, :float, precision: 8, scale: 4, default: nil
   end
 end
